# L0/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple service to get order data by its ID.",
        "title": "L0 Order Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/order/{orderID}": {
            "get": {
                "description": "Retrieves a complete order object from the cache or database using its unique ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get an order by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order Unique ID",
                        "name": "orderID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved order",
                        "schema": {
                            "$ref": "#/definitions/dto.PublicOrder"
                        }
                    },
                    "400": {
                        "description": "Invalid order ID provided",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Order with the specified ID was not found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "An internal server error occurred",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.PublicItem": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nm_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "sale": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 0
                },
                "size": {
                    "type": "string"
                },
                "total_price": {
                    "type": "integer"
                }
            }
        },
        "dto.PublicOrder": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "delivery_cost": {
                    "type": "integer"
                },
                "delivery_service": {
                    "type": "string"
                },
                "goods_total": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PublicItem"
                    }
                },
                "order_uid": {
                    "type": "string"
                }
            }
        }
    }
}